+++ src/exp/include/visitors/ExpToZ3Visitor.hh
@@ -117,7 +117,7 @@ protected:
   z3::expr _overflowCond;
 
 public:
-  std::unordered_map<std::string, std::pair<VarType, uint8_t>> _varToTypeSize;
+  std::unordered_map<std::string, std::pair<VarType, size_t>> _varToTypeSize;
 };
 
 } // namespace oden
